🦄 refactor: 删除了为了测试而引入的无用函数，放到test进行统一处理
