+++ components/embed_claw/tools/tools_web_search.c
@@ -275,26 +275,3 @@ static esp_err_t ec_tool_web_search_execute(const char *input_json, char *output
     cJSON_Delete(input);
     return err;
 }
-
-void ec_tools_web_search_format_results_for_test(const char *response_json, char *output, size_t output_size)
-{
-    cJSON *root;
-
-    if (!output || output_size == 0) {
-        return;
-    }
-
-    if (!response_json) {
-        snprintf(output, output_size, "No web results found.");
-        return;
-    }
-
-    root = cJSON_Parse(response_json);
-    if (!root) {
-        snprintf(output, output_size, "No web results found.");
-        return;
-    }
-
-    format_results_tavily(root, output, output_size);
-    cJSON_Delete(root);
-}
